1. Add support compiler >= Borland C++ 6.0 2. Fix warnings 3. Refactoring 4. Update readme 5. Update screenshots

+++ include/LDL/Events/Event.hpp
@@ -39,22 +39,22 @@ namespace LDL
 				LDL::Events::LostFocus   LostFocus;
 			};
 
-			bool IsKeyPressed(size_t key)
+			bool IsKeyPressed(uint8_t key)
 			{
 				return (Type == IsKeyboard && Keyboard.Key == key && Keyboard.State == Enums::ButtonState::Pressed);
 			}
 
-			bool IsKeyReleased(size_t key)
+			bool IsKeyReleased(uint8_t key)
 			{
 				return (Type == IsKeyboard && Keyboard.Key == key && Keyboard.State == Enums::ButtonState::Released);
 			}
 
-			bool IsMousePressed(size_t key)
+			bool IsMousePressed(uint8_t key)
 			{
 				return (Type == IsMouseClick && Mouse.Button == key && Mouse.State == Enums::ButtonState::Pressed);
 			}
 
-			bool IsMouseReleased(size_t key)
+			bool IsMouseReleased(uint8_t key)
 			{
 				return (Type == IsMouseClick && Mouse.Button == key && Mouse.State == Enums::ButtonState::Released);
 			}
