+++ source/Platforms/Windows/Graphics/MainWindow.hpp
@@ -25,7 +25,7 @@ namespace LDL
 			const Math::Vec2u& Size();
 			const Math::Vec2u& Pos();
 		private:
-			size_t ConvertKey(size_t key);
+			uint8_t ConvertKey(size_t key);
 			static LRESULT CALLBACK WndProc(HWND Hwnd, UINT Message, WPARAM WParam, LPARAM LParam);
 			LRESULT CALLBACK Handler(UINT Message, WPARAM WParam, LPARAM LParam);
 			BaseWindow _BaseWindow;
